Rename nav-bar.js -> NavBar.js
This might necessitate restarting your dev server.
