+++ src/components/CreateShoppingList/CreateShoppingList.js
@@ -2,8 +2,6 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
-import './CreateShoppingList.css';
-
 import CreateShoppingListForm from '../CreateShoppingListForm';
 
 function CreateShoppingList({ user, loggingIn, currentStore }) {
