+++ src/components/NavBar.js
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import LogOut from './auth-components/logout';
 import './component.css';
+
 export class NavBar extends React.Component {
   render() {
     let logOutButton;
